[Kernel] Phase 4 batch-2: move clean JIT operator groups into kernels.ops (RFC #29630)
Consolidated single PR for the directly-imported JIT operator groups:
- gemm: cutedsl_bf16_gemm, cutedsl_dsv3_fused_a_gemm, fp8_blockwise_gemm, fused_a_gemm
- communication: all_reduce, mp
- layernorm: fused_eh_norm, rmsnorm_hf
- speculative: ngram_corpus, resolve_future_token_ids, ngram_embedding
- mamba: transfer_mamba, inkling_sconv
- quantization: awq/gptq marlin(_repack), fp8_quantize, mxfp8, minimax_quant_ue8m0
- kvcache: hicache, kvcache, hisparse, minimax_store_kv_index, fused_fp8_qkv_kv_cache, triton_store_cache

Each operator -> sglang.kernels.ops.<group>.<op> (classified); compat shim at the
old sglang.jit_kernel.<op> path re-exports everything (removed in Phase 5). csrc
stays in sglang.jit_kernel. Invariant CLEAN; 68 kernels tests pass; lint clean.

Tangled subsystems (attention/moe/diffusion subdirs, new model/lplb/kv_canary
groups, minimax_m3 & triton splits) + the csrc/include structural finale follow.
